+++ src/lib/components/generation/FaceFixSettings.svelte
@@ -18,10 +18,10 @@
 
   const recommendedModels: RecommendedModel[] = [
     {
-      label: "YOLOv8m Face (Recommended)",
-      filename: "face_yolov8m.pt",
-      url: "https://huggingface.co/Bingsu/adetailer/resolve/main/face_yolov8m.pt",
-      sha256: "717923c19b3f4bbf5250b728f1fa6b2cb72a33aed1d236ea9caf0e21ad943e5f",
+      label: "YOLO11n Face Seg (Recommended)",
+      filename: "Anzhc Face seg 640 v4 y11n.pt",
+      url: "https://huggingface.co/Anzhc/Anzhcs_YOLOs/resolve/0319daeae9ae40752c2fb3904069cb35cc61d2ec/Anzhc%20Face%20seg%20640%20v4%20y11n.pt",
+      sha256: "1e77ad7bd349babd8a4a90478bfc965348642b63a8d95d3b43ee13db42fd0a64",
     },
     {
       label: "YOLOv8n Face (Lightweight)",